Moving code from telemetry to telemetry_utils
This code needs to be included in the header of the C++ api, unlike the
rest of the C api.
